Rename geocoding package into geolocation
