+++ app/src/main/java/com/github/onedirection/geolocation/DeviceLocationProvider.java
@@ -1,4 +1,4 @@
-package com.github.onedirection.geocoding;
+package com.github.onedirection.geolocation;
 
 import android.Manifest;
 import android.annotation.SuppressLint;
