+++ app/src/main/java/com/github/onedirection/geolocation/NominatimGeocoding.java
@@ -1,4 +1,4 @@
-package com.github.onedirection.geocoding;
+package com.github.onedirection.geolocation;
 
 import android.content.Context;
 import android.util.Log;
@@ -16,7 +16,6 @@ import org.json.JSONObject;
 
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
-import java.util.function.Function;
 
 /** OpenStreetMap's geocoding service. */
 public final class NominatimGeocoding implements GeocodingService {
