+++ app/src/test/java/com/github/onedirection/geolocation/CoordinatesTest.java
@@ -1,4 +1,4 @@
-package com.github.onedirection.geocoding;
+package com.github.onedirection.geolocation;
 
 import org.junit.Test;
